Added zsh plugins, changed xfce terminal colors, added programs to ubuntu installer, added vscode plugins installer

+++ installers/ubuntu.sh
@@ -19,30 +19,32 @@ function install_deb_from_url() {
 
 
 # Add usb3 SSD as usb-device instead of uas
-echo "options usb-storage quirks=174c:55aa:u" | sudo tee /etc/modprobe.d/usb-storage.conf
-sudo update-initramfs -u
+if [ ! -f /etc/modprobe.d/usb-storage.conf ]; then
+    echo "options usb-storage quirks=174c:55aa:u" | sudo tee /etc/modprobe.d/usb-storage.conf
+    sudo update-initramfs -u
+fi
 
 sudo apt update
 
-sudo apt install vim git xfce4-terminal byobu tmux iotop htop zsh kcalc digikam cifs-utils libdbusmenu-gtk* vlc mpv nmap sysfsutils sysstat \
+sudo apt install vim git xfce4-terminal byobu tmux iotop htop zsh kcalc digikam cifs-utils vlc mpv nmap sysfsutils sysstat \
                  network-manager-openvpn kate remmina xdotool gimp kmymoney konversation hfsprogs qdirstat parallel cdck -y
 
 sudo apt install kronometer ktimer -y
 sudo apt install xsel xclip samba kdenetwork-filesharing -y
-sudo apt install virt-manager virt-viewer openssh-server k3b calligra -y
+sudo apt install openssh-server k3b calligra -y
 
 sudo apt install exfat-fuse exfat-utils -y
 
 
 
 # Tor browser
-sudo apt install torbrowser-launcher -y
+#sudo apt install torbrowser-launcher -y
 
 
 sudo chmod 4711 /usr/bin/wodim; sudo chmod 4711 /usr/bin/cdrdao
 
 # Realtek drivers for MacBook
-sudo apt install firmware-b43-installer -y
+#sudo apt install firmware-b43-installer -y
 
 sudo apt upgrade -y
 sudo apt autoremove -y
@@ -50,8 +52,8 @@ sudo apt autoremove -y
 
 sudo chsh -s "$(which zsh)" mandy
 
-sudo snap install intellij-idea-community --classic
-sudo snap install sublime-text --classic
+sudo snap install code --classic
+bash "$DIR/code.sh"
 
 # Etcher
 echo "deb https://deb.etcher.io stable etcher" | sudo tee /etc/apt/sources.list.d/balena-etcher.list
@@ -74,18 +76,20 @@ grep -qF -- "$LINE" "$FILE" || echo "$LINE" | sudo tee "$FILE"
 
 
 # Albert
-cd /tmp
-wget -nv -O Release.key https://build.opensuse.org/projects/home:manuelschneid3r/public_key
-sudo apt-key add - < Release.key
-sudo sh -c "echo 'deb http://download.opensuse.org/repositories/home:/manuelschneid3r/xUbuntu_18.10/ /' > /etc/apt/sources.list.d/home:manuelschneid3r.list"
-sudo apt-get update
-sudo apt install albert -y
-rm Release.key
-
+which albert >/dev/null 2>&1
+if [ $? -ne 0 ]; then
+    cd /tmp
+    wget -nv -O Release.key https://build.opensuse.org/projects/home:manuelschneid3r/public_key
+    sudo apt-key add - < Release.key
+    sudo sh -c "echo 'deb http://download.opensuse.org/repositories/home:/manuelschneid3r/xUbuntu_18.10/ /' > /etc/apt/sources.list.d/home:manuelschneid3r.list"
+    sudo apt-get update
+    sudo apt install albert -y
+    rm Release.key
+fi
 
 
 # Qt development
-sudo apt install kdevelop qt5-default build-essential mesa-common-dev libglu1-mesa-dev -y
+# sudo apt install kdevelop qt5-default build-essential mesa-common-dev libglu1-mesa-dev -y
 
 # Development
 sudo apt install shellcheck -y
@@ -99,20 +103,25 @@ sudo apt-get install -y \
     curl \
     gnupg2 \
     software-properties-common
-curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
-sudo apt-key fingerprint 0EBFCD88
-sudo add-apt-repository \
-   "deb [arch=amd64] https://download.docker.com/linux/ubuntu \
-   $(lsb_release -cs) \
-   stable"
-sudo apt-get update
-sudo apt-get install -y docker-ce
-sudo usermod -aG docker "$(whoami)"  
 
+which docker >/dev/null 2>&1
+if [ $? -ne 0 ]; then
+    curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
+    sudo apt-key fingerprint 0EBFCD88
+    sudo add-apt-repository \
+    "deb [arch=amd64] https://download.docker.com/linux/ubuntu \
+    $(lsb_release -cs) \
+    stable"
+    sudo apt-get update
+    sudo apt-get install -y docker-ce
+    sudo usermod -aG docker "$(whoami)"  
+fi
 
-sudo curl -L "https://github.com/docker/compose/releases/download/1.23.1/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
-sudo chmod +x /usr/local/bin/docker-compose
-
+which docker-compose >/dev/null 2>&1
+if [ $? -ne 0 ]; then
+    sudo curl -L "https://github.com/docker/compose/releases/download/1.23.1/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
+    sudo chmod +x /usr/local/bin/docker-compose
+fi
 # Locale
 sudo locale-gen nl_BE
 sudo locale-gen nl_BE.UTF-8
@@ -121,3 +130,10 @@ sudo locale-gen en_GB.UTF-8
 sudo locale-gen en_US
 sudo locale-gen en_US.UTF-8
 sudo update-locale
+
+
+which fzf >/dev/null 2>&1
+if [ $? -ne 0 ]; then
+    git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf
+    yes | ~/.fzf/install
+fi
